Restructure folder structure for opensearch indexer lambda and related files to make tests work as well as building  docker file
