+++ data_management/opensearch_indexer/Dockerfile
@@ -5,6 +5,6 @@ COPY requirements.txt ${LAMBDA_TASK_ROOT}
 
 RUN pip install -r requirements.txt
 
-COPY lambda_function.py ${LAMBDA_TASK_ROOT}
+COPY opensearch_indexer/ ${LAMBDA_TASK_ROOT}/opensearch_indexer
 
-CMD [ "lambda_function.lambda_handler" ]
+CMD [ "opensearch_indexer.lambda_function.lambda_handler" ]
