+++ data_management/opensearch_indexer/opensearch_indexer/index_file_content_and_metadata_in_opensearch.py
